move Attachments def to rdf
and fix rdf authors annotation
